moved JAWS back into CASS
JAWS had to be moved back into CASS because the LO extension for Eclipse does not seem to be able to export external class folders. After the file was moved back, I was able to export a new .oxt extension file, install it into LO, and run the deliverable 1 test.