added temporary JUnit test practicing using JSemcor
the test open a Semcor resource (the br-a01 context in the brown1 concord)
for each sentence in the context, it finds the first sense-tagged word, its sense, and surrounding context. These are all printed to the console

in future commits, this functionality should be expanded to not find the first sense-tagged word as the target. Also, this should be packaged as a class instead of a JUnit test. The class will be a resource in testing to get an example test set for the algorithms and the correct sense for the word.